# Conflicts:
#	CMakeLists.txt
#	lib/src/phy/rf/CMakeLists.txt
#	lib/src/phy/rf/rf_dev.h

+++ lib/src/phy/channel/gauss.c
@@ -1,12 +1,7 @@
-/**
+/*
+ * Copyright 2013-2020 Software Radio Systems Limited
  *
- * \section COPYRIGHT
- *
- * Copyright 2013-2015 Software Radio Systems Limited
- *
- * \section LICENSE
- *
- * This file is part of the srsLTE library.
+ * This file is part of srsLTE.
  *
  * srsLTE is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as
@@ -24,20 +19,19 @@
  *
  */
 
-
-#include <string.h>
-#include <stdlib.h>
 #include <math.h>
+#include <stdlib.h>
 
-float rand_gauss (void) {
-  float v1,v2,s;
+float rand_gauss(void)
+{
+  float v1, v2, s;
 
   do {
-    v1 = 2.0 * ((float) rand()/RAND_MAX) - 1;
-    v2 = 2.0 * ((float) rand()/RAND_MAX) - 1;
+    v1 = 2.0 * (rand() / (float)RAND_MAX) - 1;
+    v2 = 2.0 * (rand() / (float)RAND_MAX) - 1;
 
-    s = v1*v1 + v2*v2;
-  } while ( s >= 1.0 || s == 0.0);
+    s = v1 * v1 + v2 * v2;
+  } while (s >= 1.0 || s == 0.0);
 
-    return (v1*sqrt(-2.0 * log(s) / s));
+  return (v1 * sqrtf(-2.0 * log(s) / s));
 }
